cleaned up domain mod
