+++ server/src/domain/instrument_static/types.rs
@@ -3,12 +3,12 @@
 // 🧠 Internal modules
 use crate::model::{Exchange, InstrumentType};
 
-#[derive(Debug, Clone)]
-pub struct Instrument {
-    pub internal_symbol: String,
-    pub exchange_symbol: String,
-    // add more fields as needed (e.g. tick_size, contract_size, etc.)
-}
+// #[derive(Debug, Clone)]
+// pub struct Instrument {
+//     pub internal_symbol: String,
+//     pub exchange_symbol: String,
+//     // add more fields as needed (e.g. tick_size, contract_size, etc.)
+// }
 
 #[derive(Debug, thiserror::Error)]
 pub enum RefDataError {
@@ -36,14 +36,14 @@ pub enum RefDataError {
     InstrumentMismatch { request: String, expected: String },
 }
 
-#[derive(Debug, thiserror::Error)]
-pub enum SubscriptionError {
-    #[error("Unknown exchange")]
-    UnknownExchange,
-    #[error("Invalid instrument type: {0}")]
-    InvalidInstrumentType(String),
-    #[error("Invalid feed")]
-    InvalidFeed,
-    #[error("Instrument not found")]
-    InstrumentNotFound,
-}
+// #[derive(Debug, thiserror::Error)]
+// pub enum SubscriptionError {
+//     #[error("Unknown exchange")]
+//     UnknownExchange,
+//     #[error("Invalid instrument type: {0}")]
+//     InvalidInstrumentType(String),
+//     #[error("Invalid feed")]
+//     InvalidFeed,
+//     #[error("Instrument not found")]
+//     InstrumentNotFound,
+// }
