+++ server/src/domain/sub.rs
@@ -1,4 +1,4 @@
-// server/src/domain/subscription/mod.rs
+// server/src/domain/sub.rs
 
 // 🧠 Internal modules
 use crate::model::{Exchange, RequestedFeed};
